pyproject.toml
[tool.poetry]
name = "commodore-defaults-commentator"
version = "0.1.0"
description = ""
authors = ["Simon Gerber <simon.gerber@vshn.ch>"]

[tool.poetry.dependencies]
python = "^3.8"
reclass = {git = "https://github.com/projectsyn/reclass", rev = "feat/params-keep-uris"}
click = "^8.0.3"
syn-commodore = "^0.10.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
commodore-defaults-commentator = 'commodore_defaults_commentator.cli:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
